Send the running context to outbound handlers.
This allows their context handlers to use it as a parent and get canceled when the connection is closed.

This is important especially for outbound handlers since if the remote party hangs up and the connection is closed it's important for any long running tasks to get canceled that instant and have the handler then act on that.