Switch away from using http.Header.Write() due to it stripping \n
This is needed for some header strings that need to be sent to FreeSWITCH. For example the `extra-headers` header used in `sendevent NOTIFY` when providing a contact URI or the to/from URIs.

This change makes the following header possible:
extra-headers: Messages-Waiting: yes\r\nMessage-Account: example@1.1.1.1\r\nVoice-Message: 1/0\r\n

The main tradeoff here may be an increase of time building the header string. However when testing this out the difference was within microseconds and within error margins.